Refactor serwist integration to use next
