+++ .gitignore
@@ -14,6 +14,10 @@
 # production
 /build
 
+# serwist build output
+public/sw.js
+public/sw.js.map
+
 # TypeScript
 tsconfig.tsbuildinfo
