+++ app/lib/serwist-provider.ts
@@ -1,2 +1,2 @@
 "use client";
-export { SerwistProvider } from "@serwist/turbopack/react";
+export { SerwistProvider } from "@serwist/next/react";
