+++ serwist.config.js
@@ -1,0 +1,7 @@
+// @ts-check
+import { serwist } from "@serwist/next/config";
+
+export default serwist({
+  swSrc: "app/sw.ts",
+  swDest: "public/sw.js",
+});
